Fix burger panel so it stays on top of the stack
To make it easier to manage and remember (at least for me), I set the
mobile burger panel's z-index to 10 since it is like a "layer" on top of
the base canvas. The base canvas can use z-indices 1-9 to manage its
positioned elements, while the burger panel can use z-indices 11 (which
I used for the close button) up to 19. And so on.